requirements.txt
rich
google-generativeai
requests
